requirements.txt
llama_cpp_python==0.1.78
numpy
pyjson5